refactor(sdk-keys): rename Gate to SDKKeysPage, page to SDKEnvironmentKeysSettings
Per review feedback — avoid short-lived Gate naming convention.
SDKKeysPage is now the route entry point with the flag switch,
SDKEnvironmentKeysSettings is the clean new page.
