+++ frontend/web/components/pages/sdk-keys/index.ts
@@ -1,2 +1,1 @@
-export { default as SDKKeysGate } from './SDKKeysGate'
 export { default as SDKKeysPage } from './SDKKeysPage'
